Update local/RadiologyExaminationResultProstate.t.json
Changed name from "Imaging examination of a body structure" to "Imaging examination of a prostate", "Imaging examination of an anomaly - lesion" to "Imaging examination of a lesion", "Imaging examination of an anomaly - other findings" to "Imaging examination of other finding". Changed Swedish names to equivalent. Changed cardinality for "Imaging examination of a lesion" and "Imaging examination of other finding" from 0..1 to 0..*